chroma-lar/requirements.txt
h5py
matplotlib
numpy
pydantic
PyYAML
rich
scipy
tqdm
trimesh
rich
scikit-build
fire